error message on delete when a lecturer has presententations

+++ ospp-application/src/main/java/ch/fhnw/ip6/ospp/service/LecturerService.java
@@ -1,13 +1,16 @@
 package ch.fhnw.ip6.ospp.service;
 
 import ch.fhnw.ip6.ospp.model.Lecturer;
+import ch.fhnw.ip6.ospp.model.Presentation;
 import ch.fhnw.ip6.ospp.persistence.LecturerRepository;
+import ch.fhnw.ip6.ospp.persistence.PresentationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Component
@@ -15,13 +18,23 @@ import java.util.Optional;
 public class LecturerService {
 
     private final LecturerRepository lecturerRepository;
+    private final PresentationRepository presentationRepository;
 
     public Lecturer save(Lecturer lecturer) {
         return lecturerRepository.save(lecturer);
     }
 
     public void delete(Long id) {
-        lecturerRepository.deleteById(id);
+        Optional<Lecturer> lecturer = lecturerRepository.findById(id);
+        if (lecturer.isPresent()) {
+            List<Presentation> presentations = presentationRepository.findByCoachOrExpert(lecturer.get(), lecturer.get());
+            if (presentations.isEmpty()) {
+                lecturerRepository.deleteById(id);
+            } else {
+                String joinedPresentations = presentations.stream().map(p -> String.valueOf(p.getId())).collect(Collectors.joining(","));
+                throw new IllegalArgumentException(String.format("Lehrperson ist noch Präsention (%s) zugewiesen.", joinedPresentations));
+            }
+        }
     }
 
     public void deleteAll() {
